# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Royals Cricket Academy
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Registration {
  id: number;
  receiptNumber: string;
  studentName: string;
  dateOfBirth: string;
  gender: string;
  parentName: string;
  phone: string;
  email?: string;
  address: string;
  /** junior, senior, or elite */
  category: string;
  batchTiming: string;
  registrationFee: number;
  monthlyFee: number;
  paymentMode: string;
  status: string;
  registeredAt: string;
}

export interface CreateRegistrationBody {
  studentName: string;
  dateOfBirth: string;
  gender: string;
  parentName: string;
  phone: string;
  email?: string;
  address: string;
  category: string;
  batchTiming: string;
  registrationFee: number;
  monthlyFee: number;
  paymentMode: string;
}

export interface Receipt {
  id: number;
  receiptNumber: string;
  registrationId: number;
  studentName: string;
  parentName: string;
  phone: string;
  category: string;
  batchTiming: string;
  /** registration, monthly, or other */
  feeType: string;
  amount: number;
  paymentMode: string;
  paymentDate: string;
  month?: string;
  notes?: string;
  issuedAt: string;
}

export interface GenerateReceiptBody {
  registrationId: number;
  feeType: string;
  amount: number;
  paymentMode: string;
  paymentDate: string;
  month?: string;
  notes?: string;
}

export interface StatsSummary {
  totalStudents: number;
  juniorCount: number;
  seniorCount: number;
  eliteCount: number;
  totalFeesCollected: number;
  thisMonthRegistrations: number;
  activeStudents: number;
}