# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Royals Cricket Academy
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all student registrations
 */
export const ListRegistrationsResponseItem = zod.object({
  id: zod.number(),
  receiptNumber: zod.string(),
  studentName: zod.string(),
  dateOfBirth: zod.string(),
  gender: zod.string(),
  parentName: zod.string(),
  phone: zod.string(),
  email: zod.string().optional(),
  address: zod.string(),
  category: zod.string().describe("junior, senior, or elite"),
  batchTiming: zod.string(),
  registrationFee: zod.number(),
  monthlyFee: zod.number(),
  paymentMode: zod.string(),
  status: zod.string(),
  registeredAt: zod.string(),
});
export const ListRegistrationsResponse = zod.array(
  ListRegistrationsResponseItem,
);

/**
 * @summary Register a new student
 */
export const CreateRegistrationBody = zod.object({
  studentName: zod.string(),
  dateOfBirth: zod.string(),
  gender: zod.string(),
  parentName: zod.string(),
  phone: zod.string(),
  email: zod.string().optional(),
  address: zod.string(),
  category: zod.string(),
  batchTiming: zod.string(),
  registrationFee: zod.number(),
  monthlyFee: zod.number(),
  paymentMode: zod.string(),
});

/**
 * @summary Get a student registration by ID
 */
export const GetRegistrationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetRegistrationResponse = zod.object({
  id: zod.number(),
  receiptNumber: zod.string(),
  studentName: zod.string(),
  dateOfBirth: zod.string(),
  gender: zod.string(),
  parentName: zod.string(),
  phone: zod.string(),
  email: zod.string().optional(),
  address: zod.string(),
  category: zod.string().describe("junior, senior, or elite"),
  batchTiming: zod.string(),
  registrationFee: zod.number(),
  monthlyFee: zod.number(),
  paymentMode: zod.string(),
  status: zod.string(),
  registeredAt: zod.string(),
});

/**
 * @summary List all fee receipts
 */
export const ListReceiptsResponseItem = zod.object({
  id: zod.number(),
  receiptNumber: zod.string(),
  registrationId: zod.number(),
  studentName: zod.string(),
  parentName: zod.string(),
  phone: zod.string(),
  category: zod.string(),
  batchTiming: zod.string(),
  feeType: zod.string().describe("registration, monthly, or other"),
  amount: zod.number(),
  paymentMode: zod.string(),
  paymentDate: zod.string(),
  month: zod.string().optional(),
  notes: zod.string().optional(),
  issuedAt: zod.string(),
});
export const ListReceiptsResponse = zod.array(ListReceiptsResponseItem);

/**
 * @summary Get receipts for a specific registration
 */
export const GetReceiptsByRegistrationParams = zod.object({
  registrationId: zod.coerce.number(),
});

export const GetReceiptsByRegistrationResponseItem = zod.object({
  id: zod.number(),
  receiptNumber: zod.string(),
  registrationId: zod.number(),
  studentName: zod.string(),
  parentName: zod.string(),
  phone: zod.string(),
  category: zod.string(),
  batchTiming: zod.string(),
  feeType: zod.string().describe("registration, monthly, or other"),
  amount: zod.number(),
  paymentMode: zod.string(),
  paymentDate: zod.string(),
  month: zod.string().optional(),
  notes: zod.string().optional(),
  issuedAt: zod.string(),
});
export const GetReceiptsByRegistrationResponse = zod.array(
  GetReceiptsByRegistrationResponseItem,
);

/**
 * @summary Generate a fee receipt for a registration
 */
export const GenerateReceiptBody = zod.object({
  registrationId: zod.number(),
  feeType: zod.string(),
  amount: zod.number(),
  paymentMode: zod.string(),
  paymentDate: zod.string(),
  month: zod.string().optional(),
  notes: zod.string().optional(),
});

/**
 * @summary Get academy registration statistics
 */
export const GetStatsSummaryResponse = zod.object({
  totalStudents: zod.number(),
  juniorCount: zod.number(),
  seniorCount: zod.number(),
  eliteCount: zod.number(),
  totalFeesCollected: zod.number(),
  thisMonthRegistrations: zod.number(),
  activeStudents: zod.number(),
});

/**
 * @summary Get recent registrations
 */
export const GetRecentRegistrationsResponseItem = zod.object({
  id: zod.number(),
  receiptNumber: zod.string(),
  studentName: zod.string(),
  dateOfBirth: zod.string(),
  gender: zod.string(),
  parentName: zod.string(),
  phone: zod.string(),
  email: zod.string().optional(),
  address: zod.string(),
  category: zod.string().describe("junior, senior, or elite"),
  batchTiming: zod.string(),
  registrationFee: zod.number(),
  monthlyFee: zod.number(),
  paymentMode: zod.string(),
  status: zod.string(),
  registeredAt: zod.string(),
});
export const GetRecentRegistrationsResponse = zod.array(
  GetRecentRegistrationsResponseItem,
);