# lib/api-zod/src/generated/types/createRegistrationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Royals Cricket Academy
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRegistrationBody {
  studentName: string;
  dateOfBirth: string;
  gender: string;
  parentName: string;
  phone: string;
  email?: string;
  address: string;
  category: string;
  batchTiming: string;
  registrationFee: number;
  monthlyFee: number;
  paymentMode: string;
}