# lib/api-zod/src/generated/types/generateReceiptBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Royals Cricket Academy
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateReceiptBody {
  registrationId: number;
  feeType: string;
  amount: number;
  paymentMode: string;
  paymentDate: string;
  month?: string;
  notes?: string;
}