# lib/api-zod/src/generated/types/receipt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Royals Cricket Academy
 * OpenAPI spec version: 0.1.0
 */

export interface Receipt {
  id: number;
  receiptNumber: string;
  registrationId: number;
  studentName: string;
  parentName: string;
  phone: string;
  category: string;
  batchTiming: string;
  /** registration, monthly, or other */
  feeType: string;
  amount: number;
  paymentMode: string;
  paymentDate: string;
  month?: string;
  notes?: string;
  issuedAt: string;
}