# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Royals Cricket Academy
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalStudents: number;
  juniorCount: number;
  seniorCount: number;
  eliteCount: number;
  totalFeesCollected: number;
  thisMonthRegistrations: number;
  activeStudents: number;
}